requirements.txt
telethon>=1.34.0
pyTelegramBotAPI>=4.14.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
Flask
gigachat
